feat: Add Uniswap v3 connector

+++ interfaces/IConnectionPool.sol
@@ -3,6 +3,6 @@ pragma solidity ^0.8.0;
 
 interface IConnectionPool {
     function isRegisteredConnector(address _connector) external view returns (bool);
-    function getConnectionAddress(uint _connectionId) external view returns (address);
+    function getConnectionAddress(uint24 _connectionId) external view returns (address);
     function getConnectionMax() external view returns (uint);
 }
